second commit

+++ src/main/java/com/example/demo/repository/QuestionRepo.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.model.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepo extends JpaRepository<Question, Integer> {
+public interface QuestionRepo extends JpaRepository<Question, Long> {
 
 }
